requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.6.3
asgiref==3.11.0
blinker==1.9.0
cachetools==4.2.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
decouple==0.0.7
Django==5.2.10
django-cors-headers==4.9.0
django-decouple==2.1
django-unfold==0.87.0
djangorestframework==3.16.1
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
fastapi-mail==1.5.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jwt==1.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
passlib==1.7.4
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
python-telegram-bot==13.15
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.39.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.5
starlette==0.48.0
termcolor==3.1.0
tornado==6.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
